Custom validations with admission webhooks (#55)
Introduced a CustomValidator running as a standalone webhook server for validating CRD Spec prior to k8 persisting the changes.

Validation currently hinders end user from trying to change to cluster mode in a destructive way. More validations can later on be implemented.

Cert-manager is required on the cluster as a prerequisite, otherwise webhook server won't start. Running make run locally ignores the webhook server unless explicitly started with the ENV ENABLED_WEBHOOKS=true, this is set in docker ENV, in order to force the admission webhook running the container img.

Question: should one be able to run without the admission webhook? In order to save resources on the cluster, and also removing the need for the cert manager, if one is OK with skipping the validations.